requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
motor==3.5.1
pymongo==4.8.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Validation
pydantic==2.9.0
pydantic-settings==2.5.2
email-validator==2.2.0

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.2

# Notifications (mock/simulation)
# boto3==1.35.0  # For AWS SNS/SES in production
